Refactor fetching hooks to reuse helper functions
